pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "geofusion"
version = "0.1.0"
description = "Industrial Multimodal AI for CAD and Geometry Understanding"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    "pandas>=2.1.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
    "matplotlib>=3.8.0",
    "plotly>=5.18.0",
    "omegaconf>=2.3.0",
    "h5py>=3.10.0",
    "faiss-cpu>=1.7.4",
    "transformers>=4.35.0",
    "trimesh>=4.0.0",
]

[project.optional-dependencies]
pyg = [
    "torch-geometric>=2.4.0",
    "torch-scatter",
    "torch-sparse",
    "torch-cluster",
]
full = [
    "open3d>=0.17.0",
    "wandb>=0.16.0",
    "tokenizers>=0.15.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "ipykernel>=6.27.0",
    "jupyter>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["geofusion*"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
